move all files to tsx
