+++ src/features/Risk/components/BattleList.jsx
