+++ src/features/Risk/components/ResultDisplay.jsx
