+++ src/screens/Hooks/useActionHashScroll.js
@@ -1,5 +1,6 @@
 import { useEffect } from 'react';
 import { scrollToHash, getHash } from '../../shared-utils/src';
+
 export const useActionHashScroll = () => {
   useEffect(() => {
     const initHash = getHash();
